ENH: Refactor interpolation to handle pipelines and better organization
