+++ podpac/data.py
@@ -5,9 +5,9 @@ Data Public Module
 # REMINDER: update api docs (doc/source/user/api.rst) to reflect changes to this file
 
 from podpac.core.data.datasource import DataSource
-from podpac.core.data.interpolate import (
-    Interpolation, InterpolationException, interpolation_trait,
-    INTERPOLATION_METHODS, INTERPOLATION_SHORTCUTS, INTERPOLATION_DEFAULT
+from podpac.core.data.interpolation import (
+    Interpolation, InterpolationException, interpolation_trait, INTERPOLATION_DEFAULT,
+    INTERPOLATORS, INTERPOLATION_METHODS, INTERPOLATORS_DICT, INTERPOLATION_METHODS_DICT
 )
 from podpac.core.data.types import (
     Array, PyDAP, Rasterio, WCS, ReprojectedSource, S3, H5PY, CSV
