+++ podpac/interpolators.py
@@ -4,6 +4,8 @@ Interpolators Public Module
 
 # REMINDER: update api docs (doc/source/user/api.rst) to reflect changes to this file
 
-from podpac.core.data.interpolate import (
-    Interpolator, NearestNeighbor, NearestPreview, Rasterio, ScipyGrid, ScipyPoint
+
+from podpac.core.data.interpolator import Interpolator
+from podpac.core.data.interpolators import (
+    NearestNeighbor, NearestPreview, Rasterio, ScipyGrid, ScipyPoint
 )
